requirements.txt
Django>=5.2
djangorestframework
pyrebase4
firebase-admin
